requirements.txt
asgiref==3.7.2
black==23.10.0
click==8.1.7
colorama==0.4.6
coverage==7.3.2
cssbeautifier==1.14.9
Django==4.2.6
django-dotenv==1.4.2
django-htmx==1.17.0
django-jazzmin==2.6.0
django-taggit==4.0.0
djlint==1.34.0
EditorConfig==0.12.3
factory-boy==3.3.0
Faker==19.11.0
flake8==6.1.0
html-tag-names==0.1.2
html-void-elements==0.1.0
inflection==0.5.1
iniconfig==2.0.0
jsbeautifier==1.14.9
json5==0.9.14
Markdown==3.5
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
Pillow==10.1.0
platformdirs==3.11.0
pluggy==1.3.0
pycodestyle==2.11.1
pyflakes==3.1.0
pytest==7.4.2
pytest-cov==4.1.0
pytest-django==4.5.2
pytest-factoryboy==2.6.0
python-dateutil==2.8.2
PyYAML==6.0.1
regex==2023.10.3
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
